requirements.txt
Flask==2.0.1
gunicorn==20.1.0
praw==7.4.0
python-dotenv==0.19.0
